feat: Add drag and drop functionality documentation and improve ChatContext with tool execution handling
