+++ src/context/ChatContext.tsx
@@ -4,18 +4,28 @@ import React, {
   useContext,
   useEffect,
   useMemo,
+  useRef,
   useState,
 } from 'react';
+import { useAsyncFn } from 'react-use';
 import { useSessionContext } from './SessionContext';
 import { useSessionHistory } from './SessionHistoryContext';
 import { useAIService } from '../hooks/use-ai-service';
 import { useAssistantContext } from './AssistantContext';
 import { useBuiltInTools } from './BuiltInToolContext';
+import { useUnifiedMCP } from '../hooks/use-unified-mcp';
+import { useScheduler } from './SchedulerContext';
 import { createId } from '@paralleldrive/cuid2';
 import { getLogger } from '../lib/logger';
-import { Message } from '@/models/chat';
+import { Message, UIResource } from '@/models/chat';
 import { useSettings } from '../hooks/use-settings';
 import { AIServiceConfig } from '@/lib/ai-service';
+import {
+  isMCPError,
+  MCPResponse,
+  mcpResponseToString,
+  MCPResourceContent,
+} from '@/lib/mcp-types';
 
 const logger = getLogger('ChatContext');
 
@@ -25,6 +35,121 @@ interface SystemPromptExtension {
   priority: number; // Higher number = higher priority
 }
 
+interface SerializedToolResult {
+  success: boolean;
+  text?: string;
+  uiResource?: UIResource | UIResource[];
+  error?: string;
+  metadata: Record<string, unknown>;
+  toolName: string;
+  executionTime: number;
+  timestamp: string;
+}
+
+// Pure function - serializeToolResult
+const serializeToolResult = (
+  mcpResponse: MCPResponse,
+  toolName: string,
+  executionStartTime: number,
+): SerializedToolResult => {
+  const executionTime = Date.now() - executionStartTime;
+  const timestamp = new Date().toISOString();
+
+  // 표준 MCP 에러 처리
+  if (isMCPError(mcpResponse)) {
+    return {
+      success: false,
+      error: `${mcpResponse.error.message} (Code: ${mcpResponse.error.code})`,
+      metadata: {
+        toolName,
+        mcpResponseId: mcpResponse.id,
+        jsonrpc: mcpResponse.jsonrpc,
+        errorCode: mcpResponse.error.code,
+        errorData: mcpResponse.error.data,
+      },
+      toolName,
+      executionTime,
+      timestamp,
+    };
+  }
+
+  // Tool 실행 에러 처리 (isError: true)
+  if (mcpResponse.result?.isError) {
+    const errorText =
+      mcpResponse.result.content?.[0]?.type === 'text'
+        ? mcpResponse.result.content[0].text
+        : 'Tool execution failed';
+
+    return {
+      success: false,
+      error: errorText,
+      metadata: {
+        toolName,
+        mcpResponseId: mcpResponse.id,
+        jsonrpc: mcpResponse.jsonrpc,
+        isToolExecutionError: true,
+      },
+      toolName,
+      executionTime,
+      timestamp,
+    };
+  }
+
+  // 성공 케이스 - UIResource 감지 및 보존
+  let uiResources: UIResource[] = [];
+  let textContent = '';
+
+  if (mcpResponse.result?.content) {
+    // UIResource 항목들을 찾아서 수집
+    const resourceItems = mcpResponse.result.content.filter(
+      (item): item is MCPResourceContent => item.type === 'resource',
+    );
+
+    if (resourceItems.length > 0) {
+      uiResources = resourceItems.map((item) => item.resource);
+      logger.info(
+        `Found ${uiResources.length} UI resources in tool response`,
+        {
+          toolName,
+          resourceTypes: uiResources.map((r) => r.mimeType),
+        },
+      );
+    }
+
+    // 텍스트 내용 수집 - UI Resource가 있으면 요약, 없으면 전체 텍스트
+    if (uiResources.length > 0) {
+      // UIResource가 있을 때는 간단한 요약만 제공 (HTML 등 대용량 컨텐츠 제외)
+      const resourceSummary = uiResources
+        .map((r) => `${r.mimeType}${r.uri ? ` (${r.uri})` : ''}`)
+        .join(', ');
+      textContent = `UI Resources: ${resourceSummary}`;
+    } else {
+      // UIResource가 없을 때는 기존 동작 유지
+      textContent = mcpResponseToString(mcpResponse);
+    }
+  }
+
+  return {
+    success: true,
+    text: textContent,
+    uiResource:
+      uiResources.length === 1
+        ? uiResources[0]
+        : uiResources.length > 1
+          ? uiResources
+          : undefined,
+    metadata: {
+      toolName,
+      mcpResponseId: mcpResponse.id,
+      jsonrpc: mcpResponse.jsonrpc,
+      hasUIResource: uiResources.length > 0,
+    },
+    toolName,
+    executionTime,
+    timestamp,
+  };
+};
+
 interface ChatContextValue {
   submit: (messageToAdd?: Message[], agentKey?: string) => Promise<Message>;
   isLoading: boolean;
@@ -33,6 +158,7 @@ interface ChatContextValue {
     extension: Omit<SystemPromptExtension, 'id'>,
   ) => string;
   unregisterSystemPrompt: (id: string) => void;
+  cancel: () => void;
 }
 
 const ChatContext = createContext<ChatContextValue | undefined>(undefined);
@@ -47,6 +173,143 @@ interface ChatProviderProps {
 
 const DEFAULT_SYSTEM_PROMPT = 'You are a helpful assistant.';
 
+// ToolCaller 컴포넌트 - callback prop으로 tool 실행 상태 관리
+interface ToolCallerProps {
+  onToolExecutionChange: (isExecuting: boolean) => void;
+}
+
+const ToolCaller: React.FC<ToolCallerProps> = ({ onToolExecutionChange }) => {
+  const { current: currentSession } = useSessionContext();
+  const { currentAssistant } = useAssistantContext();
+  const { messages, submit } = useChatContext();
+  const { executeToolCall } = useUnifiedMCP();
+  const { schedule } = useScheduler();
+
+  const lastProcessedMessageId = useRef<string | null>(null);
+
+  const [{ loading }, execute] = useAsyncFn(
+    async (tcMessage: Message) => {
+      if (!tcMessage.tool_calls || tcMessage.tool_calls.length === 0) {
+        logger.warn('No tool calls found in message');
+        return;
+      }
+
+      // Tool 실행 시작을 부모에게 알림
+      onToolExecutionChange(true);
+
+      try {
+        logger.info('Starting tool execution batch', {
+          toolCallCount: tcMessage.tool_calls.length,
+          messageId: tcMessage.id,
+        });
+
+        const toolResults: Message[] = [];
+
+        for (const toolCall of tcMessage.tool_calls) {
+          const toolName = toolCall.function.name;
+          const executionStartTime = Date.now();
+
+          try {
+            logger.debug('Executing tool', { toolName, toolCallId: toolCall.id });
+
+            const mcpResponse = await schedule(() =>
+              executeToolCall(toolCall),
+            );
+
+            const serialized = serializeToolResult(
+              mcpResponse,
+              toolName,
+              executionStartTime,
+            );
+
+            const toolResultMessage: Message = {
+              id: createId(),
+              assistantId: currentAssistant?.id,
+              role: 'tool',
+              content:
+                serialized.text ||
+                (serialized.uiResource
+                  ? `[UIResource: ${
+                      Array.isArray(serialized.uiResource)
+                        ? serialized.uiResource.length + ' resources'
+                        : serialized.uiResource.uri || 'resource'
+                    }]`
+                  : ''),
+              uiResource: serialized.uiResource,
+              tool_call_id: toolCall.id,
+              sessionId: currentSession?.id || '',
+            };
+
+            if (!serialized.success && serialized.error) {
+              toolResultMessage.content = `Error: ${serialized.error}`;
+            }
+
+            toolResults.push(toolResultMessage);
+
+            logger.info('Tool execution completed', {
+              toolName,
+              success: serialized.success,
+              executionTime: serialized.executionTime,
+            });
+          } catch (error) {
+            logger.error('Tool execution failed', { toolName, error });
+
+            const errorMessage: Message = {
+              id: createId(),
+              assistantId: currentAssistant?.id,
+              role: 'tool',
+              content: `Error executing ${toolName}: ${error instanceof Error ? error.message : 'Unknown error'}`,
+              sessionId: currentSession?.id || '',
+              tool_call_id: toolCall.id,
+            };
+
+            toolResults.push(errorMessage);
+          }
+        }
+
+        if (toolResults.length > 0) {
+          logger.info('Submitting tool results', {
+            resultCount: toolResults.length,
+            messageId: tcMessage.id,
+          });
+
+          await submit(toolResults, currentAssistant?.id);
+        }
+      } finally {
+        // Tool 실행 완료를 부모에게 알림
+        onToolExecutionChange(false);
+      }
+    },
+    [
+      submit,
+      executeToolCall,
+      schedule,
+      currentAssistant,
+      currentSession,
+      onToolExecutionChange,
+    ],
+  );
+
+  useEffect(() => {
+    const lastMessage = messages[messages.length - 1];
+    if (
+      lastMessage &&
+      lastMessage.role === 'assistant' &&
+      lastMessage.tool_calls &&
+      lastMessage.tool_calls.length > 0 &&
+      !lastMessage.isStreaming &&
+      !loading &&
+      lastMessage.id &&
+      lastProcessedMessageId.current !== lastMessage.id
+    ) {
+      lastProcessedMessageId.current = lastMessage.id;
+      execute(lastMessage);
+    }
+  }, [messages, execute, loading]);
+
+  return null;
+};
+
 export function ChatProvider({ children }: ChatProviderProps) {
   const { messages: history, addMessage } = useSessionHistory();
   const { current: currentSession } = useSessionContext();
@@ -54,6 +317,7 @@ export function ChatProvider({ children }: ChatProviderProps) {
   const { getCurrent: getCurrentAssistant, availableTools } =
     useAssistantContext();
   const { availableTools: builtInTools } = useBuiltInTools();
+  const cancelRequestRef = useRef(false);
 
   const [streamingMessage, setStreamingMessage] = useState<Message | null>(
     null,
@@ -61,9 +325,19 @@ export function ChatProvider({ children }: ChatProviderProps) {
   const [systemPromptExtensions, setSystemPromptExtensions] = useState<
     SystemPromptExtension[]
   >([]);
+  const [isToolExecuting, setIsToolExecuting] = useState(false);
 
   // Extract window size with default fallback
   const messageWindowSize = settingValue?.windowSize ?? 20;
+
+  // Cleanup on unmount
+  useEffect(() => {
+    return () => {
+      setStreamingMessage(null);
+      setSystemPromptExtensions([]);
+      setIsToolExecuting(false);
+    };
+  }, []);
 
   // Register system prompt extension
   const registerSystemPrompt = useCallback(
@@ -101,28 +375,47 @@ export function ChatProvider({ children }: ChatProviderProps) {
       return basePrompt;
     }
 
-    // Resolve all extensions (handle both string and function types)
-    const extensionPrompts = await Promise.all(
-      systemPromptExtensions.map(async (ext) => {
-        if (typeof ext.content === 'function') {
-          return await ext.content();
-        }
-        return ext.content;
-      }),
-    );
+    try {
+      // Resolve all extensions with error handling for individual extensions
+      const extensionPrompts = await Promise.allSettled(
+        systemPromptExtensions.map(async (ext) => {
+          try {
+            if (typeof ext.content === 'function') {
+              return await ext.content();
+            }
+            return ext.content;
+          } catch (error) {
+            logger.warn('Failed to resolve system prompt extension', {
+              extensionId: ext.id,
+              error,
+            });
+            return ''; // Return empty string for failed extensions
+          }
+        }),
+      );
 
-    // Combine base prompt with extensions
-    const combinedPrompt = [basePrompt, ...extensionPrompts]
-      .filter(Boolean)
-      .join('\n\n');
+      // Extract successful results
+      const resolvedPrompts = extensionPrompts
+        .map((result) => (result.status === 'fulfilled' ? result.value : ''))
+        .filter(Boolean);
 
-    logger.debug('Built combined system prompt', {
-      baseLength: basePrompt.length,
-      extensionsCount: systemPromptExtensions.length,
-      totalLength: combinedPrompt.length,
-    });
+      // Combine base prompt with successful extensions
+      const combinedPrompt = [basePrompt, ...resolvedPrompts]
+        .filter(Boolean)
+        .join('\n\n');
 
-    return combinedPrompt;
+      logger.debug('Built combined system prompt', {
+        baseLength: basePrompt.length,
+        extensionsCount: systemPromptExtensions.length,
+        successfulExtensions: resolvedPrompts.length,
+        totalLength: combinedPrompt.length,
+      });
+
+      return combinedPrompt;
+    } catch (error) {
+      logger.error('Error building system prompt, using base prompt', { error });
+      return basePrompt;
+    }
   }, [getCurrentAssistant, systemPromptExtensions]);
 
   // AI Service configuration with tools only
@@ -137,9 +430,12 @@ export function ChatProvider({ children }: ChatProviderProps) {
 
   const {
     submit: triggerAIService,
-    isLoading,
+    isLoading: aiServiceLoading,
     response,
   } = useAIService(aiServiceConfig);
+
+  // Combined loading state: AI service loading OR tool execution
+  const isLoading = aiServiceLoading || isToolExecuting;
 
   // Combine history with streaming message, avoiding duplicates
   const messages = useMemo(() => {
@@ -201,11 +497,27 @@ export function ChatProvider({ children }: ChatProviderProps) {
 
   const submit = useCallback(
     async (messageToAdd?: Message[], agentKey?: string): Promise<Message> => {
+      logger.info("submit ", {messageToAdd});
       if (!currentSession) {
         throw new Error('No active session available for message submission');
       }
 
+      if (cancelRequestRef.current) {
+        cancelRequestRef.current = false;
+        logger.info("canceled!")
+        return {
+          id: createId(),
+          content: 'Request cancelled',
+          role: 'system',
+          sessionId: currentSession.id,
+          isStreaming: false,
+        };
+      }
+
       try {
+        // Clear any previous streaming state before starting new request
+        setStreamingMessage(null);
+        
         let messagesToSend = messages;
 
         // Process and validate new messages if provided
@@ -267,9 +579,11 @@ export function ChatProvider({ children }: ChatProviderProps) {
           // Update streaming state and persist to history
           setStreamingMessage(finalizedMessage);
           await addMessage(finalizedMessage);
+          return finalizedMessage;
         }
 
-        return aiResponse;
+        // Handle case where no response was received
+        throw new Error('No response received from AI service');
       } catch (error) {
         logger.error('Message submission failed', { error, agentKey });
         setStreamingMessage(null);
@@ -283,8 +597,13 @@ export function ChatProvider({ children }: ChatProviderProps) {
       triggerAIService,
       addMessage,
       systemPromptExtensions,
+      buildSystemPrompt,
     ],
   );
+
+  const handleCancel = useCallback(() => {
+    cancelRequestRef.current = true;
+  }, []);
 
   const value: ChatContextValue = useMemo(
     () => ({
@@ -293,11 +612,18 @@ export function ChatProvider({ children }: ChatProviderProps) {
       messages,
       registerSystemPrompt,
       unregisterSystemPrompt,
+      cancel: handleCancel
     }),
-    [messages, submit, isLoading, registerSystemPrompt, unregisterSystemPrompt],
+    [messages, submit, isLoading, registerSystemPrompt, unregisterSystemPrompt, handleCancel],
   );
 
-  return <ChatContext.Provider value={value}>{children}</ChatContext.Provider>;
+  return (
+    <ChatContext.Provider value={value}>
+      {children}
+      {/* ToolCaller를 자동으로 포함 - 사용자가 수동으로 추가할 필요 없음 */}
+      <ToolCaller onToolExecutionChange={setIsToolExecuting} />
+    </ChatContext.Provider>
+  );
 }
 
 export function useChatContext(): ChatContextValue {
