+++ src/features/chat/Chat.tsx
@@ -33,7 +33,6 @@ import React, {
 } from 'react';
 import ToolsModal from '../tools/ToolsModal';
 import MessageBubble from './MessageBubble';
-import { ToolCaller } from './ToolCaller';
 import { TimeLocationSystemPrompt } from '../prompts/TimeLocationSystemPrompt';
 // import { useWebMCPServer } from '@/context/WebMCPContext';
 // import { ContentStoreServer } from '@/lib/web-mcp/modules/content-store';
@@ -62,7 +61,6 @@ function Chat({ children }: ChatProps) {
       {/* <JailbreakSystemPrompt /> */}
       <div className="h-full w-full font-mono flex flex-col rounded-lg overflow-hidden shadow-2xl">
         {children}
-        <ToolCaller />
         <ToolsModal
           isOpen={showToolsDetail}
           onClose={() => setShowToolsDetail(false)}
